refactor: BaseNumbers, BonusNumber 클래스 삭제
BaseNumbers를 삭제한 이유: BonusNumber와 MainNumbersContainer가 BaseNumbers 추상 클래스를 상속 받지만 두 클래스가 서로를 대체할 수 없음. 즉 LSP 원칙에 위반됨.

BonusNumber를 삭제한 이유: LottoNumber로 대체 가능
