+++ src/main/java/lotto/base_numbers/MainNumbersContainer.java
@@ -1,16 +1,57 @@
 package lotto.base_numbers;
 
+import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
+import java.util.stream.Collectors;
 import lotto.LottoNumber;
 
-public class MainNumbersContainer extends BaseNumbers {
+public class MainNumbersContainer {
 
     private static final int REQUIRED_QUANTITY = 6;
 
+    private List<LottoNumber> numbers;
+
     public MainNumbersContainer(List<LottoNumber> numbers) {
-        super(numbers);
+        Collections.sort(numbers);
+        this.numbers = List.copyOf(numbers);
         validate(numbers);
+    }
+
+    public boolean contain(LottoNumber lottoNumber) {
+        return numbers.contains(lottoNumber);
+    }
+
+    public int countContain(MainNumbersContainer other) {
+        int count = 0;
+        for (LottoNumber n : other.numbers) {
+            if (this.numbers.contains(n)) count++;
+        }
+        return count;
+    }
+
+    @Override
+    public String toString() {
+        String str = numbers.stream()
+                .map(LottoNumber::toString)
+                .collect(Collectors.joining(", "));
+        return "[" + str + "]";
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (!(o instanceof MainNumbersContainer other)) {
+            return false;
+        }
+        if (this.numbers.size() != other.numbers.size()) {
+            return false;
+        }
+        return this.numbers.equals(other.numbers);
+    }
+
+    @Override
+    public int hashCode() {
+        return numbers.hashCode();
     }
 
     private void validate(List<LottoNumber> numbers) {
