+++ src/test/java/lotto/LottoTest.java
@@ -4,7 +4,6 @@ import static lotto.TestLottoNumber.*;
 import static lotto.TestLottoNumbers.*;
 
 import java.util.ArrayList;
-import lotto.base_numbers.BonusNumber;
 import lotto.base_numbers.MainNumbersContainer;
 import lotto.draw_numbers.DrawNumbers;
 import lotto.draw_numbers.DrawNumbersBuilder;
@@ -35,10 +34,9 @@ class LottoTest {
     @Test
     void 로또_번호에서_당첨_번호가_정상적으로_카운트됩니다() {
         MainNumbersContainer winningNumbers = new MainNumbersContainer(new ArrayList<>(BASIC_NUMBERS));
-        BonusNumber bonusNumber = new BonusNumber(new ArrayList<>(List.of(SEVEN)));
         DrawNumbers drawNumbers = DrawNumbersBuilder.builder()
                 .winningNumbers(winningNumbers)
-                .bonusNumber(bonusNumber)
+                .bonusNumber(SEVEN)
                 .build();
         Lotto lotto = new Lotto(new MainNumbersContainer(new ArrayList<>(BASIC_NUMBERS)));
         Assertions.assertEquals(lotto.countWinningNumbers(drawNumbers), 6);
@@ -47,10 +45,9 @@ class LottoTest {
     @Test
     void 로또_번호에서_보너스_번호가_정상적으로_인식됩니다() {
         MainNumbersContainer winningNumbers = new MainNumbersContainer(new ArrayList<>(BASIC_NUMBERS));
-        BonusNumber bonusNumber = new BonusNumber(new ArrayList<>(List.of(SEVEN)));
         DrawNumbers drawNumbers = DrawNumbersBuilder.builder()
                 .winningNumbers(winningNumbers)
-                .bonusNumber(bonusNumber)
+                .bonusNumber(SEVEN)
                 .build();
         Lotto lotto = new Lotto(new MainNumbersContainer(new ArrayList<>(List.of(ONE, TWO, THREE, FOUR, FIVE, SEVEN))));
         Assertions.assertTrue(lotto.containBonusNumber(drawNumbers));
